Rename: purchase에서 product로 변경, action, reducer 수정
