+++ client/src/components/Layout/UserAuth.js
@@ -1,7 +1,7 @@
 import { Fragment } from "react";
 import { useDispatch } from "react-redux";
 import { Link, useNavigate } from "react-router-dom";
-import { logout } from "../../_actions/user_action";
+import { logout } from "../../_actions/userAction";
 
 import classes from "./UserAuth.module.css";
