+++ client/src/components/Product/Slide.js
