+++ client/src/hoc/auth.js
@@ -8,7 +8,7 @@
 
 import React, { useEffect } from "react";
 import { useDispatch } from "react-redux";
-import { auth } from "../_actions/user_action";
+import { auth } from "../_actions/userAction";
 import { useNavigate } from "react-router-dom";
 
 // eslint-disable-next-line import/no-anonymous-default-export
